requirements.txt
psycopg2-binary==2.9.3
pytest==6.2.5
neo4j==4.4.1
pymongo==4.3.3
redis==4.3.5
cassandra-driver==3.25.0
